Add detection of rst2man to build
This adds support for finding rst2man and will then generate the fvwm.1 man
pahe.  If rst2man isn't found on the system, /bin/true is used in place of
rst2man and a warning is emitted.

The warning will probably become fatal over time, when the work on the man
page is complete.