Add detection of asciidoctor to build
This adds support for finding asciidoctor and will then generate the fvwm.1 man
page.  If asciidoctor isn't found on the system, /bin/true is used in place of
it and a warning is emitted.

The warning will probably become fatal over time, when the work on the man
page is complete.