Remove hard-coded USEDECOR macro
The USEDECOR macro was presumably introduced a long time ago in FVWM's history
when the concept of window decoration was first introduced.  This has been set
to "1" for a long time, making the code unnecessarily conditional based on a
value which could never change.

This is a small (albeit very small) step for removing the decor code and
eventually replacing it with something more flexible.