Make PNG support mandatory
Traditionally, FVWM has always had the option of not linking to any image
renderer at all.

But, the entire world has moved on, and support for images is becoming more
important, especially with third-party applications.  Indeed, FVWM itself will
have a new default configuration soon which will make use of PNGs.  It's a
good idea to support this out of the box.

Choosing the default image library is not easy -- and although there's a sway
towards SVGs (which FVWM supports), PNGs are quite common still, so use that.