Add default-config to fvwm
It's been a long-time coming, but fvwm out-of-the-box now has an uptodate and
maintainable configuration file which will hopefully serve as a good basis for
users to customise, should they not have one of their own to start with.

A few tiny backend changes (by Thomas Adam) were required to plumb the config
into the existing infrastructure.

All credit for this goes to Jaimos Skriletz for this work, with some nice
feedback from the fvwm community at-large (mostly on #fvwm on freenode).