FIX: EdgeCommand is sometimes not triggered.
If focus has been stolen by some unmanaged window (this for
example happens with GTK-menus), EdgeCommands are not triggered.
The reason for this is the fact that the focus is restored
before the check is performed that determines if a command is
defined for the given edge, causing the wrong window to be
used for that check.